requirements.txt
boto==2.49.0
boto3==1.16.25
data-science-types==0.2.20
pytest==6.1.2
pytest-mock==3.3.1
python-decouple==3.3
requests
typing==3.6.6
numpy==1.16.0
moto==1.3.16
pandas==0.24.2

